Restore livelogging code for killing unique monsters
This was evidently quietly removed during one of the 3.7 merges with
none the wiser. Put it back, and also refactor it to have a little less
of a Wizard-vs-everyone-else case.

This also reinstates the system of only logging some repeated kills of
the same unique monster. But the wording is improved from the former (5
times) that it used to be.