Add easter eggs which only generate on Easter
These aren't a new item type, they're just a fruitname which is assigned
to slime molds generated during level creation when it is Easter.
Wishing and horns of plenty and other means of creating food items after
the level generates will give the fruit the player's current fruitname,
as usual.