Add "gold" and "gilded" random wand appearances
This doesn't add any actual new objects, it just sticks them onto the
back of the back of the wands list so that these appearances will be
randomly shuffled in with the rest of the wands, as was already the case
for a few wand types.

The main purpose of doing this is so that there can be wands made out of
gold that are stealable by leprechauns.