Object materials: always render the material on some types of armor
These are the leather armor pieces that had "leather" stripped from
their names, which has caused some confusion among players since it's
not immediately obvious what a "light armor" is actually made of, for
instance. This commit always forces the material to be shown for such
armor pieces.

Based on a commit from EvilHack.