Flag some additional monsters as herbivorous and carnivorous
This is to partially address the oddity that if you polymorph into a
form that is neither herbivorous nor carnivorous, you are inediate and
have no nutritional requirements at all. This makes sense for some
monsters (undead for instance) but doesn't for creatures that probably
should have some sort of a metabolism.

Nymphs are herbivorous. Salamanders, Kops, and the types of impish
creatures that can leave corpses are carnivorous.

The original plan called for making bees herbivorous too, but then I
realized that the ramifications of that would include things like pet
bees eating lichen corpses, and bees do not work that way.