Prevent ghosts from being renamed
I was always confused as to why ghosts were able to be renamed, as this
changes *who they are the ghost of*. There are a number of interesting
things you can do with ghosts with a certain name, but all of them fall
apart if you can just rename the ghost to whatever you want. So, make it
impossible to do that.