Port L's Cursed Wand Backfire Patch
This was a small 3.4.3-era patch that aimed to further disincentivize
zapping a cursed wand (because the 1% chance of it exploding is rare
enough that they're considered safe to use in most situations). Under
this patch, a cursed directional wand will instead zap you (as if you
had aimed it at yourself) 10% of the time.