Wand of opening extracts items from statues
With statues able to be made of non-diggable materials, it was
impossible to get at anything inside them if one happened to be copper
or gold. This is a long-planned addition that makes an opening beam
(only from the wand, not from the spell of knock) spill the contents of
any statue it hits. It also works for statues in your inventory if you
zap yourself.

The best function for handling objects spilling out of a container is
tipcontainer(), but that assumed that the hero is tipping a container
on their own space; thus, most of this commit is refactoring it to also
handle objects being spilled out of a container that isn't on the hero's
space, and controlling messages appropriately.