Change S_engraving's default ASCII glyph to ~
People complained about confusing engravings with amulets, especially
since amulets can now generate as other materials whose colors conflict
with engravings (is a gray " a carved engraving or a silver amulet?)

In order to try and avoid this, engravings now render as ~, in the same
colors they used to render as. This does cause a color conflict between
dust engravings and long worm tails; hopefully that's a rare enough case
not to matter much. (In any case, an unknown ~ glyph can be farlooked to
see what it actually is). Players that want to keep it as " can define
SYMBOLS=S_engraving:" in their config file.

Because purple ~ (vibrating square) and bright purple ~ (graffiti) could
be confused, change the color of the vibrating square glyph to yellow
(which is what it was in 3.6.0, and also stands out a fair bit better
than purple).