Fix: segfault when potion of acid exploded in water
This dates back to when I first made water_damage_chain able to damage a
fixed number of objects (commit 06f9b31). The loop that iterates over
items being damaged did not account for the fact that items (such as
potions of acid) can be destroyed and deleted by the water, filling them
with a garbage nobj pointer that would be used after free.