Fix: zombified monsters kept their mextra structs
This led to behavior such as a zombified shopkeeper remaining in their
shop to tend it, or a zombified priest remaining in their temple and
continuing to intone "Pilgrin, you enter a sacred place!" - very
unzombielike behavior.

I have made it so that zombification wipes all the mextra structs and
flags indicating the presence of mextra structs. I debated changing how
zombification works entirely, by having the monster die as normal and
then spawning a zombie in its place, but ultimately figured that it
would cause a lot of weirdness (a corpse and inventory would be dropped,
the monster generation would have to be deferred until at least the
remove_monster of the old monster so that it didn't get enexto'd, etc),
and cleaning up this weirdness would be a very messy job.

This also adds code to preserve the name of the monster and attach that
to the new zombie.