Fix: clear trap ammo when creating a trap
Had some bug reports from the 'deleting trap containing ammo'
impossible, specifically with weird ttyp/otyp combinations such as
STATUE_TRAP and ROCK. This appears to be coming from the fact that
maketrap() can overwrite a preexisting trap with a new one, and this
code didn't touch the trap's ammo field, making errors like the one
above possble if a statue trap generated on and overwrote a falling rock
trap.

This first extends set_trap_ammo to delete the trap's ammo if it's
passed a null object pointer, and then uses that to clear any ammo in
the trap during the trap fields reset in maketrap().

Also, add impossible sanity checks to set_trap_ammo and
remove_trap_ammo.