Object materials: metal armor can't go below 1 AC (unenchanted)
Noticed this when I found some copper gauntlets of power that provided
no AC, which is rather silly; +0 metal gloves ought to be better than
nothing at all. Adjust the AC-by-material formula to not deduct points
below 1 when the object is made of metal.

This will affect most helms, kicking boots, dwarvish boots (if gold),
and gauntlets of power. There are no base-1-AC cloaks, suits, or shirts
which can be metal.