Fix: infinite loop trying to find a space for starting room graffiti
An assumption made in 3b0c12b was that there would always be an
available ROOM space in the first room of the dungeon. With theme rooms,
this is technically no longer the case; it's possible that the only
valid room to place the stairs in is the "Ice room" or something.
Granted, this is unlikely enough to be mostly latent, but it's
theoretically possible, especially due to a vanilla bug that allows
stairs to generate in themed rooms.

Use find_okay_roompos, which already has built-in limits for the loop.