Fix: irregular rooms' walls were not part of the room
This was leading to colored walls problems with special themed rooms
which were irregular. Walls of ordinary rooms count as part of the room,
and irregular rooms should be no different.

This also makes doors on the room edge part of the room, which affected
special room entry messages and filling.

This may be a latent vanilla bug (that all irregular rooms' walls get
marked as SHARED instead of their own room). Not sure this is the
correct solution, but it seems to do the job.