Add new ebones mextra structure
This structure is designed to contain extra information about a monster
that was created from a player dying and leaving bones, such as a ghost.
It contains basic fields (role, race, gender, alignment), plus several
fields that have no immediate use but may become useful later, such as
luck, crowning and the demigod flag. The values are currently set during
bones creation but not read anywhere.

This also required defining a NUM_RACES constant in order to iterate
through races to get the index of a player's race.

This is a save-breaking commit.