Fix: "split room" subroom could have width/height of 0
This computes an x or y coordinate at which to "draw the wall"; I think
this code was an artifact of when I was doing this by actually drawing a
wall rather than making a subroom. It had 1 subtracted from it in the
width/height computation, which meant the subroom could have 0 width or
height.