Refactor and replace remove_trap_ammo with new deltrap_with_ammo
I wasn't happy with how remove_trap_ammo was structured: it
unconditionally deleted the trap, almost as a side effect, and you had
the choice of either placing the ammo on the ground or burying it, when
there are other things to do with it. This commit throws it out and
replaces it with a new deltrap_with_ammo function. This acts identically
to deltrap for traps that don't have ammo, and for those that do it can
either delete the ammo, place it, bury it, or return the object chain to
the caller.

The impossible logic for when deltrap is called with ammo is adjusted.
The general idea going forward is that deltraps that trigger that
impossible can be drop-in replaced with deltrap_with_ammo as needed.