Wearing cursed gear prints a weld message and auto-identifies
This is aimed at making worn equipment behave the same way as weapons
do, and avoiding annoying behaviors such as taking off
non-beatitude-known gear in order to determine whether it has a curse.
This was information the player already had available, but it was
bothersome to get the information.

xNetHack already contained a message for worn cursed gear becoming
uncursed "relaxing slightly", so this provides the counterpart to that.

This behavior also happens when non-cursed, worn gear becomes cursed, as
from a curse items spell. The principle is the same: if the player is
wearing cursed gear, that should be known to them. (Note: wearing
something non-cursed and autocursing like a dunce cap will not print a
weld message immediately when you put it on, but then the autocurse will
trigger and it will print the message.)

The one item which is exempt from this message is the amulet of change;
a weld message would be pointless since it selfdestructs.