Dying on a corridor or grass leaves a grave in bones
This is from NetHack Fourk (at least the corridor part; I added grass to
it). make_grave was being cautious about where a grave could be placed,
and only permitted placing a grave on a ROOM or GRAVE space, possibly
out of caution for not putting a grave in a vault guard's fake corridor.
I simply used the existing in_fcorridor to block that one instance of
make_grave from occurring if the player did die in the fake corridor.
According to jonadab, all the other instances of make_grave already take
necessary precautions.

Originally I threw in ice as a valid grave location too, but later
decided to pull it out since it would be weird for a person dying on ice
that then melts to create a little grave islet which is walkable. If
only the game had separate furniture and terrain layers...