Fix: starting in polyinit mode could drop objects at (0,0)
Due to the polymon() call happening before you get placed on the
upstairs of the first level. Move it to after you get placed there. (No
messages are printed for this polymon() call, meaning you can now start
the game standing on a pile of the weapons and armor you dropped by
polymorphing, if the chosen polyinit form is one that would make you
drop things.)