Fix: riding steed into polytrap called deltrap twice
This change comes with one side effect: unchanging does not protect the
steed when you ride it into a polymorph trap. This is because unlike in
vanilla, you can get unchanging from sources such as arising from the
dead or polyinit mode, which clearly shouldn't protect the steed from
polymorphs. I considered suggesting to the devteam that they make the
same change, but it's a weaker argument when unchanging comes only from
the amulet.

This also fixes a latent unrelated bug where a steed riding into a
polytrap would remove the polytrap regardless of whether it actually got
affected.