Fix: three bugs with ghosts displaying as I incorrectly
The first bug was that if a ghost made its "appearance" attack and you
couldn't spot it after it became uninvisible, it would be mapped as an I
regardless. I vacillated on whether this should keep doing this, but
print a message to justify the I appearing from nowhere; but ultimately
decided that that would result in a lot of message spam in places around
graveyards. So there is now a ghost exception to the general mhitu case
that an attacking monster you can't spot automatically gets an I.

The second bug had much the same effect, deriving from a different place
in the code.

The third bug was that when a ghost became invisible, it could then
call newsym and render as an I even if you couldn't see it in the first
place. (This may have been latent, as calling newsym on a location you
can't see won't usually do anything.)