Use technical rather than colloquial orientation in options display
This was based on a pull request to SpliceHack by Michael Meyer, but I
ended up figuring out that I didn't need most of its changes because I
had accounted for them one way or the other when initially implementing
orientations. (xNetHack doesn't currently have any need for
distinguishing between initial and current orientation; nothing can
change it in-game and even if there were something, it would probably be
fairly pointless to distinguish between initial and final orientation in
xlogfiles or something.)

What this does flip is the term shown in the options menu; it used to
render as "straight/gay/bi", now it will render as
"heterosexual/homosexual/bisexual".