Running and traveling no longer push boulders
This was a perennial annoyance for players, in which they could run down
a hallway and push a boulder they didn't know was there into a blocked
position, or push random boulders out of the way while traveling. This
commit makes it so that only a deliberate, single-step move can move a
boulder.

Adapted from SpliceHack pull request #83 with a change to fix a major
bug present in that pull request.