Object materials: dmgval "weapon vs monster type" bonuses on all objects
This was a set of four bonuses that applied only to weapons, ball&chain,
and gems being used to hit an monster. It seems to me that these bonuses
should apply regardless of object class.

- Blessed objects' d4 damage versus undead-types. Doesn't make sense
  that you'd get this bonus by wielding a blessed rock and hitting
  things with it, but not other types of blessed objects.
- Axes versus wooden monsters. Only applies to weapons anyway.
- Light-emitting artifact versus a light-hater. Only applies to
  Sunsword, but there's no specific reason why this should be restricted
  to weapons.
- And most importantly, material hatred. Given that people can do things
  like wield silver wands against silver haters, there's no reason that
  needs its own special cases to be handled elsewhere.

This commit is in preparation for a future commit, specifically a
refactor of special_dmgval().