Object materials: overhaul special_dmgval()
This replaces the clunky guts of this function with more straightforward
logic, without actually changing the function signature. Additionally,
it now checks for special damage from the attacker's body in all cases,
not just player-as-attacker.

The core of this damage function now just calls dmgval() rather than
duplicating a subset of its code.

If multiple things are coming in contact with the defender and applying
a damage bonus, each one rolls its own damage bonus, then the one with
the highest roll will be applied and be returned to the caller. The
ultimate effect is that this function is biased to return a higher
value when there are multiple things causing damage, which makes sense.

This is in preparation for an upcoming object materials commit.