Remove expulsion and allow questing with the leader dead, but at a price
The main point of this feature is to remove the possibility for
unwinnable games via getting expelled from the quest permanently, prior
to questing, with no way back in. Much of the code that does this is
drawn from UnNetHack. A side effect is that a significant barrier (of
gaining a bunch of levels) in pacifist games is removed as well.

The player still gets banished from the quest upon doing things that
would get them banished forever previously. But the quest portal isn't
sealed off, and the leader becomes angry. The leader remains peaceful
until you approach them after being "exiled".

What other variants have done in allowing leader death to open up the
quest is to add measures to make the quest leader harder to kill,
usually at least by making them immune to death rays, and often further
buffs to their combat skills. I don't really like this approach for
three reasons:
1) It doesn't make any flavor sense that quest leaders are inherently
   death-magic-resistant.
2) It seems like it closes off one route of trivializing the battle
   without closing off other routes, not that that is a good solution
   either since it just becomes an arms race between the player and the
   developer.
3) The quest leader is not always supposed to be a powerful warrior
   (e.g. Hippocrates is supposed to be sick and frail; making him a
   beefy fighter would undercut that flavor). It allows for less nuanced
   quests.

So instead of that, I went for a softer disincentive to enter the quest
by way of dead leader, in the form of permanent bad luck penalties.
These exist whether or not you personally kill the leader, and make it
hard for you to keep your luck up, especially in the later game. If you
do kill the leader personally, you are saddled with massive luck and god
anger penalties, because the existing large alignment record penalty was
ineffective on its own. I'm not totally happy with these, since they can
be rectified by altar farming and have no long-term consequences
afterward, but it didn't seem like there was a save-stable way to track
the status of the player killing the leader.

Also, direct kills of quest guardians got smaller (but still
significant) luck and god anger penalties.