Conflict negates Elbereth and scare monster
This is from FIQHack, and the general idea is to prevent the player from
using conflict to kill most of the monsters in an area while being
mostly or totally safe themselves. Note that it does not affect scaring
via music, reading the scroll of scare monster, or other means, and does
not block vampires from being scared of the player standing on an altar.