Cursed amulet of life saving only works half of the time
This is a venerable feature originating in SporkHack and copied in a
bunch of variants. (I modeled the implementation on UnNetHack, though
changed the messages and raised the chance of this happening to what
SporkHack has.)

The reason for this is to add another incentive to keep your amulet
non-cursed. Previously, if you were committed to using a life saving
amulet for the rest of the game, it didn't matter if it got cursed, you
weren't going to remove it anyway. Now it matters.