Lifesaving livelog shows the would-be killer
"Averted death" is common and boring in most circumstances. Why not show
the cause of death the player got lifesaved from? It might draw more
attention that way.

I tried to replace the "averted death" message completely, but wasn't
able to make the resulting clause grammatically correct in all cases
(e.g. some deaths such as "dissolved in molten lava" should be prepended
with "was", and some such as "died of starvation" should not, and then
the deaths that have no killer prefix would all need to be accounted for
separately...) so I just had it print the formatted death reason in
parentheses.