Make Nazgul ungenocideable
This is a small component of a feature from EvilHack (sorry, they're not
going to start riding fell beasts any time soon).

I doubt anyone really genocides Nazgul very often, but it seems to me
that a species with a low limited birth cap shouldn't be able to be
wiped out by genocide.