Fix: teleport region in Bar-goal didn't work correctly
I misunderstood how des.teleport_region() is intended to work; in this
case, what I wanted was for an exclusion zone that both a) prevents the
hero from levelporting and being able to end up next to Thoth Amon and
b) prevents the hero from arriving on the level and immediately
teleporting in next to him. Because teleport regions currently only
allow you to define two rectangles and the physical barriers on the map
are irregular, it will be possible to walk into a place where you can
then teleport directly to Thoth Amon.

The teleport region controlling your landing now places the hero fairly
near the stairs. I also added a second teleport region to the level,
which prevents them from teleporting straight to Thoth Amon after they
leave that small area.