Unify monnam calls for livelogs that should ignore hallucination
Noticed that my fix in f1f7cc2 only extended to preventing uniques from
being livelogged as bogusmons when hallucinating, and the livelog for
killing a player remnant monster in bones still had this problem.

This defines a macro, livelog_mon_nam(), which wraps the appropriate
x_monnam call, and has both livelog cases use that.