Make number of polymorphed items in #conduct visible outside wizard mode
It was already visible in wizard mode, but uncloak it for all games.
By exposing it in #conduct, it will carry through to the dumplog, and
thus it will be possible to tell after the fact whether a game that
broke polypileless conduct did a minimal amount of polymorphing items
(e.g. not really caring about the conduct and not performing any mass
polypiling) and a game that did do mass polypiling.

I don't really see any possibility of abuse of showing this number, but
if some exploit turns up, this can always be walled behind a game-over
conditional.