Put grass back in the Bar-goal level
This was sidelined while it was impossible in vanilla to do
replace_terrain with a selection. Now it's possible, and can use the
gradient selection to do this. I adjusted the gradient bounds slightly
to get a little more grass nearer the town, and additionally prevented
grass from spawning in the buildings.