Undead iron-haters are unaffected by iron
Someone pointed out it was weird that elf zombies and mummies still
recoiled from the touch of iron as if they could feel pain. That was a
good point, so I removed iron hatred for undead creatures (which, as
far as I know, only affects elf undead but it would also affect, say,
zombie nymphs if those were ever added).