Extend Angels' resistances to couatls and ki-rin
This is based on a Fourk commit which extended this to just couatls (on
the basis that all angelic beings should have similar resistances), but
I was surprised to see that ki-rin had no resistances at all, so I gave
them the same set.

As in Fourk, the one exception is that couatls don't get cold
resistance, because they're reptilian.

Unlike in Fourk, there's no MR_MAGM, and I didn't want to create one or
add them to resists_mgc (the basis for Angels having that is that they
have a magic missile attack).