Make mindless monsters and werewolves sleep resistant
From NetHack Fourk, although in Fourk there are more monsters made sleep
resistant (most demons, the Wizard) that I declined to give the
resistance to because allowing the player to sleep them can result in
some interesting strategies.

I initially was going to do this by sticking MR_SLEEP flags on every
affected monster, but then decided I can't think of a case in the future
where I'd add a mindless monster and want it to be vulnerable to sleep,
so this augments the definition of resists_sleep itself, making the
MR_SLEEP flag on mindless monsters redundant.