Named weapons can pull their names from the top ten list
This is a feature from SpliceHack, creating weapons with names like
"Sabre of Luxidream". I also pulled in some of its additions to
wpn_names.

This diverges in two ways: first, the name from the top ten list is
always capitalized because I thought names like "Mace of benjamin" would
look weird, and instead of a flat 1/20 chance for receiving a top ten
name, the chance is 1/(25 - 2*enchantment). Meaning that a highly
enchanted weapon is more likely to be named for a player.

I had to add a tt_name() function like Splice did in order to get a
random name from the topten list without either trying to use the
toptenentry struct outside of topten.c (a huge pain) or using tt_oname
and letting the corpsenm of the weapon be clobbered (problematic if
corpsenm on weapons ever becomes relevant).