Fix: obj materials: another two missing cases for material hatred
Touching a returning weapon flying back at you never existed as a silver
case in vanilla (probably because Mjollnir and aklyses aren't silver in
vanilla) so it got missed by the initial material hatred implementation.

This covers both the case of you catching the returning weapon (as
usual, wearing gloves will negate material damage in this case), and the
case where it flies back and hits you.