Fix: object materials: uwep shouldn't be retouched if wearing gloves
Wielding a new weapon already contains a check for wearing gloves which
makes retouch_object not even be called (which is NOT made redundant by
this fix - retouch_object must be called BEFORE it actually gets wielded
in case it's a bane artifact that can't be handled, and this fix checks
for a weapon which is already wielded).

However, the last commit added a new call to retouch_object that
specifically touches the wielded weapon, and this was causing material
damage even if wearing gloves. I added a case in retouch_object itself
that blocks taking material damage (not artifact damage!) from your
weapon if you are wearing gloves.