Fix: obj materials: postmortem searmsg call for artifact instakills
The initial bug here was with a message sequence such as "Vorpal Blade
decapitates the elf!  The elf recoils from the touch of iron!", first
noticed in EvilHack. The initial proposed fix was to add a DEADMONSTER
check to searmsg, but then I noticed that this removed the searing
message from (usually non-artifact) sequences such as "Your silver
dagger sears the vampire's flesh!  You kill the vampire!" which is
problematic because it implies that the kill occurred without any help
from bonus material damage.

I tried to figure out ways of solving this minimally, but couldn't get
past the problem that artifact_hit did not give its caller any clarity
on whether it had killed the target with a special effect or merely
inflicted damage to the target that happened to be fatal. So the
solution ended up being refactoring artifact_hit to return more info, in
the form of flags that indicate whether it printed any special message
at all and whether it printed a message indicating an instakill. Then
the callers of artifact_hit can pick up on this in order to suppress the
searmsg call if an instakill message specifically has already been
printed.