Buff potion vapor effects
This magnifies the effects of inhaling full healing, confusion,
sleeping, speed, and blindness potions to a greater extent than they
used to be. In large part this is because monsters throw some of the
offensive potions at the player fairly often, and the effect only
lasting for 1-5 turns means they don't get much out of doing that.

Hallucination used to do nothing at all; now it causes hallucination,
unless the player happens to be one of a small set of
hallucination-oriented monsters.  This part is from EvilHack.

I also decided that just getting a whiff of booze vapors shouldn't be
enough to cause any confusion, especially since the duration of
confusion from confusion vapors was just increased. Now it prints its
own message and has no effect.