Allow monsters to displace peaceful monsters out of their way
This exists mainly because it's not fair the player is the only one
capable of displacing peacefuls, because it lets them interpose a
peaceful between them and an enemy. If in a narrow corridor or maze,
they're totally safe from any melee attacks of that enemy.

This refactors mm_displacement to reorganize the rules around when and
how monsters are allowed to displace other monsters, so that there can
be an early positive exit from the function after some deny checks and
before other deny checks. That positive exit consists of a case where
the monster moving out of the way is peaceful and non-tame, enabling
monster peaceful displacement.

I wasn't sure whether to allow smaller hostile monsters to move larger
peaceful monsters out of the way. For now, they cannot move larger
monsters out of the way; it's less fair this way since the player can
still displace peacefuls larger than them, but it doesn't cause
weirdness like a newt moving a giant out of the way.

Note: this does NOT allow pets to displace peaceful monsters out of the
way, because they follow their own logic. I think there may be a vanilla
bug here, because I can't even seem to get a displacer beast to displace
any other monsters.