Fire command will automatically select a wielded autoreturning weapon
... If the hero has nothing quivered and autoquiver finds nothing to
quiver as a backup, that is. So, if you have an aklys wielded and a stack
of darts quivered, the f command will fire the darts, and when they run
out the f command will throw the aklys.

The point is to make it less painful to use autoreturning weapons (granted,
it only cuts it from t + [letter] to f, but f is familiar for repeated
throws, and a 50% reduction in keypresses is great by any metric).

I dislike how autoquiver will skip over the aklys for a knife or stack
of arrows that also happens to be in the inventory and then try to throw
them, but I guess that's why I don't usually play with autoquiver on.

I also would rather like to have this print a message the first time you
ready your wielded weapon to throw (indicating your quiver has run out),
but I can't see a way to do that which doesn't break saves.