Unchanging pauses rather than cancels sliming
This is an interesting feature I've wanted to do for a while and it
turned out to be easier than expected to implement. Logically, all
unchanging does is to prevent your form from changing. If your body
happens to have partially turned to slime, it won't magically cure it
back to normal, it'll just put the slime in stasis. I also added
messages for putting on and taking off an amulet of unchanging while
being slimed (and since they're unambiguous, they identify the amulet if
you get the message.)

Other forms of curing sliming aren't affected. Also, I had to rewrite
the main timer countdown code a bit due to it being poorly structured
for the addition of something that pauses a timer.