Fix: corruption and polyinit mode made hero immune to slime
This was due to it all using Unchanging simply and directly. Now there
is a distinction given by a new function can_slime_with_unchanging(),
which checks to see if your only source of Unchanging is the sort given
by corruption and polyinit mode.

The logic works but isn't perfect; if the hero ever becomes able to get
permanent intrinsic unchanging from some other source, it'll be treated
the same as polyinit mode is now, but I can't totally fix that without a
savebreak. Not that there are any plans for offering permanent intrinsic
unchanging.

The annoying thing is that all Unchanging checks for starting sliming
now need to also call can_slime_with_unchanging.