Fix: monsters could wear objects made of a hated material
This was largely not noticeable since monsters don't pick up objects made
of hated materials and don't generate with them either. But in case a
soldier gets polymorphed into an elf-lord or something, it will no
longer want to wear iron armor. (Note that cases like that don't inflict
material damage on the monster, because there's no retouch_object
equivalent for them.)