Fix: monsters' amulets of life saving never failed
I completely missed this when implementing amulet of life saving failure
for the player. However, this exposed a few bugs: sometimes mlifesaver()
was used to mean "monster will definitely be lifesaved". This required a
little refactoring to fix; in particular, whether a cursed amulet of
life saving will fail CANNOT be non-deterministic, otherwise code that
tries to make decisions based on whether the monster will be lifesaved
will be wrong.

To address this, I added a function that unifies the logic for checking
whether an amulet of lifesaving will fail. It is now deterministic based
on a hash of the object id (I don't see any possible leaks or exploits
here because whether or not the amulet fails, it's used up). The places
that call mlifesaver to figure out whether the monster will get
lifesaved can now also call this function to see whether they'll
actually be lifesaved.