Thiefstones can be re-keyed
One of the common pieces of feedback I've received about thiefstones is
that they're nice to have if you can get one on an accessible level, but
the problem is the ones a player finds tend to have generated on levels
that are tedious to return to later in the game, like somewhere in the
Mines. Specifically, players noted that it would be nice if you could
make a thiefstone be keyed to a level of your choice.

The first decision was that when a player re-keys a thiefstone, it
should key to a specific spot they want rather than its default behavior
of scanning the level for vaults, closets, and chests, on the basis that
this would require the player to hunt for the keyed location a second
time and/or have to learn the thiefstone searching algorithm. Also, it
would still be possible to control the keyed location on a level lacking
vaults and closets... by ensuring there was exactly one container on the
level at the desired location. Not fun to set up. Thus, the thiefstone
keys to where the player is standing.

I knew I wanted to do this via dipping the thiefstone into some potion.
Candidates were gain ability, restore ability, confusion and
hallucination. The latter two were nice because they would create a use
for a rarely used potion, but at the same time they made it a less
significant decision for the player to use one on re-keying a
thiefstone, since the player would be using up a potion that would
likely go to junk anyways. Gain ability didn't fit quite as thematically
as restore ability, especially after deciding this would only work on
cancelled thiefstones, so I selected restore ability.

Why should this only work on cancelled thiefstones? Because it pushes
the point at which the player can have a thiefstone keyed to wherever
they want later into the game. In order to get this, the player needs a
combination of four things: a thiefstone, a curse-removal method, a
cancellation method and a potion of restore ability. Hopefully in most
games this will happen somewhat into the midgame.